backend/requirements.txt
# Flask REST API
Flask>=3.0.0
Flask-CORS>=4.0.0
gunicorn>=21.2.0

# Machine Learning - Core
scikit-learn>=1.4.0
numpy>=1.26.0
joblib>=1.3.0
xgboost>=2.0.0
imbalanced-learn>=0.11.0

# ML Explainability
shap>=0.43.0

# NLP & Language Detection
langdetect>=1.0.9

# Universal Input Processing
requests>=2.31.0
beautifulsoup4>=4.12.0
pdfplumber>=0.10.0
Pillow>=10.0.0
pytesseract>=0.3.10

# Database
supabase>=2.3.0
psycopg2-binary>=2.9.0

# Data Processing
pandas>=2.1.0
python-dotenv>=1.0.0
pytz==2023.3.post1
python-dateutil==2.8.2
